+++ app/react/App/scss/layout/_header.scss
@@ -299,6 +299,10 @@ header {
       background: var(--c-bg-menu-mobile);
     }
   }
+
+  @media (min-width: 1024px) {
+    max-width: 75%;
+  }
 }
 
 .mobile-menuActions {
@@ -311,10 +315,19 @@ header {
   list-style: none;
   padding: 0;
   margin: 0;
+
+  @media (min-width: 1024px) {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    @if $theme == light {
+      color: $c-black;
+    } @else {
+      color: $c-white;
+    }
+  }
 }
 
 .menuNav-item {
-  position: relative;
   padding: 0;
   width: 100%;
   padding-bottom: 5px;
